+++ packages/dara-components/changelog.md
@@ -2,10 +2,29 @@
 title: Changelog
 ---
 
-## NEXT
+## 1.8.2
+
+-   Fix type of `default_legends` in `CausalGraphViewer` and `VisualEdgeEncoder` to reflect that it does not accept `None`. 
+-   Fixed an issue where empty `Chat` component caused an error due to having undefined length.
+
+## 1.8.1
 
 -   `Chat` component now shows the user who has written a message
 -   `Chat` component now displays its messages content as `Markdown`.
+-   It is now possible to add a callback to your configuration for when a message is sent in `Chat` component, an example can be found below:
+
+```python
+from dara.core.configuration import ConfigurationBuilder
+from dara.components.smart.chat import NewMessageBody, ChatConfig
+
+
+def example_callback(payload: NewMessageBody):
+    print('New message received!')
+    print(payload)
+
+config = ConfigurationBuilder()
+config.add_configuration(ChatConfig(on_new_message=example_callback))
+```
 
 ## 1.8.0